+++ src/main/java/ru/saiev/epictest/services/CountService.java
@@ -1,15 +1,13 @@
 package ru.saiev.epictest.services;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
+
 import ru.saiev.epictest.dto.CounterDto;
 import ru.saiev.epictest.entities.Count;
-import ru.saiev.epictest.repositories.CountRepository;
 
 
 public interface CountService {
 
-    Count getCountById(Long id);
+    Count getCountById(Integer id);
 
     Count incrementCountById(CounterDto counterDto);
